move some type and function to parent supplier
