+++ src/containers/Supplier/Table/Components/DeleteComp.tsx
@@ -3,7 +3,7 @@ import { confirmAlert } from 'react-confirm-alert';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-confirm-alert/src/react-confirm-alert.css';
 import 'react-toastify/dist/ReactToastify.css';
-import { Supplier } from "../TableSupplier";
+import { Supplier } from "../../Supplier";
 
 type Props = {
   supplier: Supplier;
